update to matrixssl 3.3

+++ alexb/lib/matrixssl/contrib/matrixssl/cipherSuite.c
@@ -1,12 +1,15 @@
 /*
  *	cipherSuite.c
- *	Release $Name: MATRIXSSL-3-2-2-OPEN $
+ *	Release $Name: MATRIXSSL-3-3-0-OPEN $
  *
  *	Wrappers for the various cipher suites.
  *	Enable specific suites at compile time in matrixsslConfig.h
  */
 /*
- *	Copyright (c) PeerSec Networks, 2002-2011. All Rights Reserved.
+ *	Copyright (c) AuthenTec, Inc. 2011-2012
+ *	Copyright (c) PeerSec Networks, 2002-2011
+ *	All Rights Reserved
+ *
  *	The latest version of this code is available at http://www.matrixssl.org
  *
  *	This software is open source; you can redistribute it and/or modify
@@ -16,8 +19,8 @@
  *
  *	This General Public License does NOT permit incorporating this software 
  *	into proprietary programs.  If you are unable to comply with the GPL, a 
- *	commercial license for this software may be purchased from PeerSec Networks
- *	at http://www.peersec.com
+ *	commercial license for this software may be purchased from AuthenTec at
+ *	http://www.authentec.com/Products/EmbeddedSecurity/SecurityToolkits.aspx
  *	
  *	This program is distributed in WITHOUT ANY WARRANTY; without even the 
  *	implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
@@ -589,7 +592,7 @@ sslCipherSpec_t *sslGetCipherSpec(ssl_t *ssl, uint32 id)
 
 	i = 0;
 	do {
-		if ((int32)supportedCiphers[i].ident == id) {
+		if (supportedCiphers[i].ident == id) {
 #ifdef USE_SERVER_SIDE_SSL		
 /*
 			Globally disabled?
